Use the Current Level Value When Input Null
I noticed that one of the dashboard integrations is only sending two of the three values in it's color command event and it was causing an unhandled exception. Adding a null replacement value using the current level so that the driver is able to complete the color change without error.